refactor: save 로직 변경
- TheaterRoom이 seat의 생명주기를 관리할 수 있도록 의도
- Transactional 추가
