+++ src/test/java/prgrms/marco/be02marbox/domain/theater/service/TheaterRoomServiceTest.java
@@ -3,9 +3,10 @@ package prgrms.marco.be02marbox.domain.theater.service;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
-import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 import javax.persistence.EntityNotFoundException;
 
@@ -20,12 +21,12 @@ import org.springframework.context.annotation.Import;
 import prgrms.marco.be02marbox.domain.theater.Seat;
 import prgrms.marco.be02marbox.domain.theater.Theater;
 import prgrms.marco.be02marbox.domain.theater.TheaterRoom;
-import prgrms.marco.be02marbox.domain.theater.converter.SeatConverter;
-import prgrms.marco.be02marbox.domain.theater.dto.SeatRequestDto;
-import prgrms.marco.be02marbox.domain.theater.dto.TheaterRoomRequestDto;
+import prgrms.marco.be02marbox.domain.theater.dto.RequestCreateSeat;
+import prgrms.marco.be02marbox.domain.theater.dto.RequestCreateTheaterRoom;
 import prgrms.marco.be02marbox.domain.theater.repository.SeatRepository;
 import prgrms.marco.be02marbox.domain.theater.repository.TheaterRepository;
 import prgrms.marco.be02marbox.domain.theater.repository.TheaterRoomRepository;
+import prgrms.marco.be02marbox.domain.theater.service.utils.SeatConverter;
 
 @DataJpaTest
 @Import({TheaterRoomService.class, SeatConverter.class})
@@ -41,25 +42,26 @@ class TheaterRoomServiceTest {
 	SeatRepository seatRepository;
 
 	private Theater theater = new Theater();
-	private List<SeatRequestDto> seatRequestDtos = new ArrayList<>();
+	private Set<RequestCreateSeat> requestCreateSeats = new HashSet<>();
 
 	@BeforeEach
 	void init() {
-		seatRequestDtos.add(new SeatRequestDto(0, 0));
+		requestCreateSeats.add(new RequestCreateSeat(0, 0));
 		theaterRepository.save(theater);
 	}
 
 	@AfterEach
 	void clear() {
-		seatRequestDtos.clear();
+		requestCreateSeats.clear();
 	}
 
 	@Test
 	@DisplayName("새로운 상영관을 추가할 수 있다.")
 	void testSave() {
-		TheaterRoomRequestDto requestDto = new TheaterRoomRequestDto(theater.getId(), "A관", seatRequestDtos);
-		int expectTotalCount = seatRequestDtos.size();
-		Long savedId = theaterRoomService.save(requestDto);
+		RequestCreateTheaterRoom requestCreateTheaterRoom = new RequestCreateTheaterRoom(theater.getId(), "A관",
+			requestCreateSeats);
+		int expectTotalCount = requestCreateSeats.size();
+		Long savedId = theaterRoomService.save(requestCreateTheaterRoom);
 
 		Optional<TheaterRoom> findTheaterRoom = theaterRoomRepository.findById(savedId);
 		List<Seat> findSeats = seatRepository.findByTheaterRoomId(savedId);
@@ -73,16 +75,17 @@ class TheaterRoomServiceTest {
 	@Test
 	@DisplayName("하나의 상영관에는 동일한 row, col을 가진 좌석이 존재할 수 없다.")
 	void testSave_distinct() {
-		seatRequestDtos.add(new SeatRequestDto(0, 1));
-		seatRequestDtos.add(new SeatRequestDto(0, 2));
-		seatRequestDtos.add(new SeatRequestDto(0, 1));
-		seatRequestDtos.add(new SeatRequestDto(0, 2));
-		seatRequestDtos.add(new SeatRequestDto(0, 1));
-		seatRequestDtos.add(new SeatRequestDto(0, 2));
+		requestCreateSeats.add(new RequestCreateSeat(0, 1));
+		requestCreateSeats.add(new RequestCreateSeat(0, 2));
+		requestCreateSeats.add(new RequestCreateSeat(0, 1));
+		requestCreateSeats.add(new RequestCreateSeat(0, 2));
+		requestCreateSeats.add(new RequestCreateSeat(0, 1));
+		requestCreateSeats.add(new RequestCreateSeat(0, 2));
 
-		TheaterRoomRequestDto requestDto = new TheaterRoomRequestDto(theater.getId(), "A관", seatRequestDtos);
+		RequestCreateTheaterRoom requestCreateTheaterRoom = new RequestCreateTheaterRoom(theater.getId(), "A관",
+			requestCreateSeats);
 		int expectTotalCount = 3;
-		Long savedId = theaterRoomService.save(requestDto);
+		Long savedId = theaterRoomService.save(requestCreateTheaterRoom);
 		List<Seat> findSeats = seatRepository.findByTheaterRoomId(savedId);
 
 		assertThat(findSeats).hasSize(expectTotalCount);
@@ -91,8 +94,9 @@ class TheaterRoomServiceTest {
 	@Test
 	@DisplayName("존재하지 않는 Theater의 요청은 EntityNotFoundException 발생")
 	void testSave_entityNotFoundException() {
-		TheaterRoomRequestDto requestDto = new TheaterRoomRequestDto(-1L, "A관", seatRequestDtos);
+		RequestCreateTheaterRoom requestCreateTheaterRoom = new RequestCreateTheaterRoom(-1L, "A관", requestCreateSeats);
 
-		assertThrows(EntityNotFoundException.class, () -> theaterRoomService.save(requestDto));
+		assertThrows(EntityNotFoundException.class, () -> theaterRoomService.save(requestCreateTheaterRoom));
 	}
 }
+
